pyproject.toml
[project]
name = "discord-mcp-alert"
version = "0.1.0"
description = ""
authors = [
    {name = "VinylStage",email = "mcpe9869@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "mcp (>=1.25.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
